psycotest was done
