+++ modulo1/lab1/.vscode/settings.json
@@ -1,3 +1,2 @@
 {
-  "python.pythonPath": "/home/serendipita/anaconda3/envs/intro/bin/python"
 }
