+++ modulo1/lab1/app.py
@@ -1,55 +1,26 @@
-from rich.console import Console
-from rich.table import Table
+import psycopg2
 
-console = Console()
+# Connect to an existing database
 
-test_data = [
-    {"jsonrpc": "2.0", "method": "sum", "params": [None, 1, 2, 4, False, True], "id": "1",},
-    {"jsonrpc": "2.0", "method": "notify_hello", "params": [7]},
-    {"jsonrpc": "2.0", "method": "subtract", "params": [42, 23], "id": "2"},
-]
+conn = psycopg2.connect(dbname='insights')
 
-def test_log():
-    enabled = False
-    context = {
-        "foo": "bar",
-    }
-    movies = ["Deadpool", "Rise of the Skywalker"]
-    console.log("Hello from", console, "!")
-    console.log(test_data, log_locals=True)
+cur = conn.cursor()
+cur.execute("SELECT * FROM clients;")
+version = cur.fetchone()
+print(version)
 
-
-test_log()
-
-
-def foo(a):
-  """This operator would be useful other things"""
-  print(a)
-
-print (foo.__doc__)
+# with conn:
+#     cur = conn.cursor()
+#     cur.execute("SELECT * FROM clients")
+    
+#     row = cur.fetchone()
+#     print(row)
+#     #print(f"{row[0]} {row[1]} {row[2]}")
+    
+#     #rows = cur.fetchall()
+#     #for row in rows:
+#     #    print(f"{row[0]} {row[1]} {row[2]}")
 
 
-console = Console()
-
-table = Table(show_header=True, header_style="bold magenta")
-table.add_column("Date", style="dim", width=12)
-table.add_column("Title")
-table.add_column("Production Budget", justify="right")
-table.add_column("Box Office", justify="right")
-table.add_row(
-    "Dev 20, 2019", "Star Wars: The Rise of Skywalker", "$275,000,000", "$375,126,118"
-)
-table.add_row(
-    "May 25, 2018",
-    "[red]Solo[/red]: A Star Wars Story",
-    "$275,000,000",
-    "$393,151,347",
-)
-table.add_row(
-    "Dec 15, 2017",
-    "Star Wars Ep. VIII: The Last Jedi",
-    "$262,000,000",
-    "[bold]$1,332,539,889[/bold]",
-)
-
-console.print(table)
+# x = cur.fetchone()[0]
+# print (x, type(x), repr(x))
